Crate glutton_westend_runtime

Glutton Westend Runtime

The purpose of the Glutton parachain is to do stress testing on the Kusama network. This runtime targets the Westend runtime to allow development separate to the Kusama runtime.

There may be multiple instances of the Glutton parachain deployed and connected to its parent relay chain.

These parachains are not holding any real value. Their purpose is to stress test the network.

Governance

Glutton defers its governance (namely, its Root origin), to its Relay Chain parent, Kusama (or Westend for development purposes).

XCM

Since the main goal of Glutton is solely stress testing, the parachain will only be able receive XCM messages from the Relay Chain via DMP. This way the Glutton parachains will be able to listen for upgrades that are coming from the Relay chain.

Modules

api

weights

xcm_config

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

ConstBool

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

EitherOfDiverse

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

IsInVec

Trivial utility for implementing Contains/OrderedMembership with a Vec.

MessageQueueServiceWeight

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

RelayOrigin

ReservedDmpWeight

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

Version

Weight

Enums

DispatchClass

A generalized group of dispatch types.

Everything

A Contains implementation that contains every value.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

build_config

Build GenesisConfig from a JSON blob not using any defaults and store it in the storage. For more info refer to sp_genesis_builder::GenesisBuilder::build_config.

create_default_config

Get the default GenesisConfig as a JSON blob. For more info refer to sp_genesis_builder::GenesisBuilder::create_default_config

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Glutton

GluttonConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

MessageQueue

Nonce

Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.